Unify the App.cfc name.
I'm trying to change as little as possible across each version of the application. This way, there's less to remember to modify from version to version; and, it means we'll use less memory (not that this is big issue in this demo). Mostly, it's so I don't have to remember to change version names. We're not caching anything in the `application` scope, so CFC references won't be an issue.