Add "load more" button to the contact list.
In addition to the prev/next pagination links (I'm leaving those in place for now) I'm adding a "load more" button to the bottom of the contacts list. This will append the next page of rows to the bottom of the existing table by selecting the `tr` elements out of the HTTP response. This way, the user doesn't lose the current page context.