Add button disabling during form processing.
This adds the `hx-disabled-elt` to disable the submission buttons during form processing. Now that I have `hx-sync` on the `body` element, the _implicit_ `hx-sync` on every element will no longer handle double-clicks on form buttons. As such, the `hx-disabled-elt` can step in to disable the buttons after the first click to pick up the slack.