Refactor search UI to use "focus scroll".
It turns out, HTMX will maintain the focus of an input across `hx-swap` operations as long as the currently-focused input has a consistent `id` on it. This is a **huge benefit** and should greatly simplify some of the swapping that has to be done within interactive interfaces. It allows me to take a much more brute-force approach since I don't have to (necessarily) worry about clobbering the user's current interaction.

There's probably subtle stuff that I'm not taking into account just yet; but, so far, this feels really good.