Add bulk delete of contacts.
This adds a checkbox to each table row, allowing several contacts to all be selected for a single bulk-delete operation. Unlike the book, which fully HTMX'ifies this step, I've opted to use a full (but still `hx-boost`ed) redirect to a specialized end-point. This just felt like an easier option.